update CR requet
